+++ puzzle.hpp
@@ -52,7 +52,7 @@ public:
     std::string type;
     int color = 0;
     int count = 0;
-    uint16_t polyshape = 0;
+    uint32_t polyshape = 0;
     std::string dir;
     int nega = NEGA_NONE;  // Added negation type
 };
@@ -75,7 +75,7 @@ public:
     bool validate();
     bool placeShapesRecursively(const std::vector<std::pair<int, int>>& positions, 
                               std::vector<std::vector<int>>& grid,
-                              const std::vector<uint16_t>& shapes,
+                              const std::vector<uint32_t>& shapes,
                               const std::vector<std::pair<int, int>>& region,
                               size_t shapeIndex = 0);
